refactor(native-select): updated component structure
