+++ packages/react/src/components/native-select/index.ts
@@ -1,0 +1,15 @@
+export {
+  NativeOption,
+  NativeOptionGroup,
+  NativeSelectRoot,
+} from "./native-select"
+export type {
+  NativeOptionGroupProps,
+  NativeOptionProps,
+  NativeSelectItem,
+  NativeSelectRootProps,
+} from "./native-select"
+
+export * from "./native-select.style"
+
+export * as NativeSelect from "./namespace"
