+++ packages/react/src/components/native-select/namespace.ts
@@ -1,0 +1,11 @@
+export {
+  NativeOption as Option,
+  NativeOptionGroup as OptionGroup,
+  NativeSelectRoot as Root,
+} from "./native-select"
+export type {
+  NativeSelectItem as Item,
+  NativeOptionGroupProps as OptionGroupProps,
+  NativeOptionProps as OptionProps,
+  NativeSelectRootProps as RootProps,
+} from "./native-select"
