+++ packages/react/src/index.ts
@@ -69,7 +69,7 @@ export * from "./components/mark"
 // export * from "./components/menu"
 export * from "./components/modal"
 export * from "./components/motion"
-// export * from "./components/native-select"
+export * from "./components/native-select"
 export * from "./components/native-table"
 // export * from "./components/notice"
 export * from "./components/number-input"
