--- packages/react/src/legacy-components/native-select/index.ts
@@ -1,7 +1,0 @@
-export { NativeOption, NativeOptionGroup, NativeSelect } from "./native-select"
-export type {
-  NativeOptionGroupProps,
-  NativeOptionProps,
-  NativeSelectItem,
-  NativeSelectProps,
-} from "./native-select"
